+++ tests/testthat/test-class_local.R
@@ -27,40 +27,6 @@ tar_test("local_init()$run() stores correct values", {
   expect_equal(out, c(1L, 1L, 10L, 20L))
 })
 
-tar_test("local_init(pipeline)$run() retains correct memory", {
-  pipeline <- pipeline_init(
-    list(
-      target_init(name = "data1", expr = quote(seq_len(10))),
-      target_init(name = "data2", expr = quote(seq_len(20))),
-      target_init(name = "min1", expr = quote(min(data1))),
-      target_init(name = "min2", expr = quote(min(data2))),
-      target_init(name = "max1", expr = quote(max(data1))),
-      target_init(name = "max2", expr = quote(max(data2))),
-      target_init(name = "mins", expr = quote(c(min1, min2))),
-      target_init(name = "maxes", expr = quote(c(max1, max2))),
-      target_init(name = "all", expr = quote(c(stop(123), mins, maxes)))
-    )
-  )
-  expect_error(
-    local_init(pipeline)$run(),
-    class = "condition_run"
-  )
-  names <- paste0(rep(c("data", "min", "max"), each = 2), seq_len(2))
-  for (name in c(names, c("mins", "maxes"))) {
-    expect_equal(
-      pipeline_get_target(pipeline, name)$cache$targets$names,
-      character(0)
-    )
-  }
-  for (name in c("mins", "maxes", "all")) {
-    expect_false(is.null(pipeline_get_target(pipeline, name)$value))
-  }
-  expect_equal(
-    pipeline_get_target(pipeline, "all")$cache$targets$names,
-    character(0)
-  )
-})
-
 tar_test("all targets unloaded at end", {
   x <- target_init("x", quote(1), memory = "persistent")
   y <- target_init("y", quote(1), memory = "transient")